Add `arity` property to selectors.
Compute the arity of the selector with the maximum parameter leangth
from the dependencies list. By comparing the `length` property of every
function primitive dependency and the computed `arity` property of every
selector dependency, the user can easily check if a selector has a
dependency that requries multiple parameters. This enables utilities to
dynamically change the behavior or memoization strategy for selectors
that may cause unnecessary recomputation when utilized by more than one
component.